+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 
 
 ## 2026-04-18
+- `21:44` -- **Added** `apps/desktopProbe/src/server/logger.ts`
 - `19:44` -- **Updated** `apps/backend/supabase/functions/_shared/logger.ts`
 - `17:42` -- **Updated** `apps/desktopProbe/src/server/logger.ts`
 - `17:37` -- **Updated** `apps/desktopProbe/forge.config.ts`